backend/skills_library/info_collection/structure_damage_alterations_recognition/requirements.txt
anthropic>=0.40.0
dashscope>=1.19.0
pdf2image>=1.17.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
